+++ app/js/modes.js
@@ -20,9 +20,8 @@
 "use strict"
 
 const modes = {
-    "normal": {"fg": "#ddd"},
+    "normal": {},
     "insert": {
-        "fg": "#3f3",
         "onLeave": newMode => {
             if (TABS.currentPage().getAttribute("dom-ready")) {
                 TABS.currentPage().send("action", "blur")
@@ -34,7 +33,6 @@ const modes = {
         }
     },
     "command": {
-        "fg": "#f33",
         "onEnter": () => {
             document.getElementById("url").value = ""
             COMMANDHISTORY.resetPosition()
@@ -42,13 +40,11 @@ const modes = {
         "onLeave": () => SUGGEST.emptySuggestions()
     },
     "search": {
-        "fg": "#ff3",
         "onEnter": () => {
             document.getElementById("url").value = ""
         }
     },
     "explore": {
-        "fg": "#3ff",
         "onEnter": () => {
             TABS.updateUrl(TABS.currentPage(), true)
             document.getElementById("url").select()
@@ -57,7 +53,6 @@ const modes = {
         "onLeave": () => SUGGEST.emptySuggestions()
     },
     "follow": {
-        "fg": "#f3f",
         "onLeave": newMode => {
             FOLLOW.cancelFollow()
             if (!["visual", "pointer"].includes(newMode)) {
@@ -66,8 +61,6 @@ const modes = {
         }
     },
     "pointer": {
-        "fg": "#777",
-        "bg": "#fff",
         "onLeave": newMode => {
             if (!["visual", "follow"].includes(newMode)) {
                 POINTER.releaseKeys()
@@ -78,8 +71,6 @@ const modes = {
         }
     },
     "visual": {
-        "fg": "#000",
-        "bg": "#3af",
         "onLeave": newMode => {
             if (!["pointer", "follow"].includes(newMode)) {
                 POINTER.releaseKeys()