+++ app/js/preloads/privacy.js
@@ -24,8 +24,8 @@ const path = require("path")
 const message = "The page has requested to view a list of all media devices."
     + " You can allow or deny this below, and choose if you want the list to"
     + " include the names (labels) of the media device in the response."
-    + " For help and options, see ':h permissionmediadevices',"
-    + " ':h permissionsallowed' and ':h permissionsblocked'."
+    + " For help and options, see ':h permissionmediadevices', ':h permissions"
+    + "allowed', ':h permissionsasked' and ':h permissionsblocked'."
 const settingsFile = path.join(
     ipcRenderer.sendSync("appdata-path"), "webviewsettings")
 
@@ -33,23 +33,12 @@ const privacyFixes = (w = window) => {
     try {
         // Hide device labels from the list of media devices
         const enumerate = w.navigator.mediaDevices.enumerateDevices
-        const mediaDeviceList = async (action, notify = false) => {
-            if (notify) {
-                ipcRenderer.sendToHost("notify",
-                    `${notify} ${action}ed 'mediadevices' at `
-                    + `'${w.location.href}' based on `
-                    + "'permissionmediadevices'", "perm")
-            }
+        const mediaDeviceList = async action => {
             if (action === "block") {
                 throw new DOMException("Permission denied", "NotAllowedError")
             }
             const devices = await enumerate.call(w.navigator.mediaDevices)
             if (action === "allowfull") {
-                if (!notify) {
-                    ipcRenderer.sendToHost("notify",
-                        `Manually allowfulled 'mediadevices' at `
-                        + `'${w.location.href}'`, "perm")
-                }
                 return devices
             }
             return devices.map(({deviceId, groupId, kind}) => ({
@@ -57,35 +46,33 @@ const privacyFixes = (w = window) => {
             }))
         }
         w.navigator.mediaDevices.enumerateDevices = async () => {
-            let action = "ask"
+            let setting = "ask"
             let settings = {}
             try {
                 settings = JSON.parse(fs.readFileSync(settingsFile).toString())
             } catch (e) {
                 // No webview settings configured, assuming the default "ask"
             }
-            for (const type of ["block", "allow"]) {
+            setting = settings.permissionmediadevices || setting
+            if (!["block", "ask", "allow", "allowfull"].includes(setting)) {
+                setting = "ask"
+            }
+            let settingRule = ""
+            for (const type of ["ask", "block", "allow"]) {
                 for (const r of settings[`permissions${type}ed`]?.split(",")) {
-                    if (!r.trim()) {
+                    if (!r.trim() || settingRule) {
                         continue
                     }
                     const [match, ...names] = r.split("~")
                     if (names.find(p => p.endsWith("mediadevices"))) {
                         if (w.location.href.match(match)) {
-                            ipcRenderer.sendToHost("notify",
-                                `Automatic rule for 'mediadevices' `
-                                + `activated at '${w.location.href}' `
-                                + `which was ${type}ed`, "perm")
-                            return mediaDeviceList(type)
+                            settingRule = type
                         }
                     }
                 }
             }
-            action = settings.permissionmediadevices
-            if (!["block", "ask", "allow", "allowfull"].includes(action)) {
-                action = "ask"
-            }
-            if (action === "ask") {
+            setting = settingRule || setting
+            if (setting === "ask") {
                 let url = w.location.href
                 if (url.length > 100) {
                     url = url.replace(/.{50}/g, "$&\n")
@@ -101,16 +88,39 @@ const privacyFixes = (w = window) => {
                 })
                 if (ask.response === 0) {
                     if (ask.checkboxChecked) {
+                        ipcRenderer.sendToHost("notify",
+                            `Manually allowfulled 'mediadevices' at `
+                            + `'${w.location.href}'`, "perm")
                         return mediaDeviceList("allowfull")
                     }
-                    action = "allow"
+                    setting = "allow"
                 }
                 if (ask.response === 1) {
-                    action = "block"
+                    setting = "block"
                 }
-                return mediaDeviceList(action, "Manually")
+                if (settingRule) {
+                    ipcRenderer.sendToHost("notify",
+                        `Ask rule for 'mediadevices' activated at '`
+                        + `${w.location.href}' which was `
+                        + `${setting}ed by user`, "perm")
+                } else {
+                    ipcRenderer.sendToHost("notify",
+                        `Manually ${setting}ed 'mediadevices' at `
+                        + `'${w.location.href}'`, "perm")
+                }
+                return mediaDeviceList(setting)
             }
-            return mediaDeviceList(action, "Globally")
+            if (settingRule) {
+                ipcRenderer.sendToHost("notify",
+                    `Automatic rule for 'mediadevices' activated at `
+                    + `'${w.location.href}' which was ${setting}ed`, "perm")
+            } else {
+                ipcRenderer.sendToHost("notify",
+                    `Globally ${setting}ed 'mediadevices' at `
+                    + `'${w.location.href}' based on `
+                    + "'permissionmediadevices'", "perm")
+            }
+            return mediaDeviceList(setting)
         }
         // Enable screensharing functionality linked to permissiondisplaycapture
         w.navigator.mediaDevices.getDisplayMedia = () => customDisplayMedia(w)
@@ -179,35 +189,68 @@ module.exports = {privacyFixes}
 // https://github.com/electron/electron/issues/16513#issuecomment-602070250
 const displayCaptureStyling = `html, body {overflow: hidden !important;}
 .desktop-capturer-selection {
-  position: fixed;top: 0;left: 0;width: 100%;height: 100vh;
-  background: #0005;color: #fff;z-index: 10000000;
-  display: flex;align-items: center;justify-content: center;
+    position: fixed;top: 0;left: 0;width: 100%;height: 100vh;
+    background: %SHADE%;color: %FG%;z-index: 10000000;text-align: center;
+    display: flex;align-items: center;justify-content: center;
 }
 .desktop-capturer-selection__scroller {
-  width: 100%;max-height: 100vh;overflow-y: auto;
+    width: 100%;max-height: 100vh;overflow-y: auto;
+}
+.desktop-capturer-selection__close {
+    position: fixed;font-size: 40px;top: 10px;right: 10px;width: 40px;
+    height: 40px;cursor: pointer;
 }
 .desktop-capturer-selection__list {
-  max-width: calc(100% - 100px);margin: 50px;padding: 0;display: flex;
-  list-style: none;overflow: hidden;justify-content: center;flex-wrap: wrap;
+    max-width: calc(100% - 100px);margin: 50px;padding: 0;display: flex;
+    list-style: none;overflow: hidden;justify-content: center;flex-wrap: wrap;
 }
 .desktop-capturer-selection__item {display: flex;margin: 10px;}
 .desktop-capturer-selection__btn {
-  display: flex;flex-direction: column;margin: 0;border: 0;width: 160px;
-  background: #333;color: #fff;align-items: center;
+    display: flex;flex-direction: column;margin: 0;border: 0;width: 160px;
+    background: %BG%;color: %FG%;align-items: center;height: auto;
+    font-size: %FONTSIZE%px;font-weight: bold;cursor: pointer;
 }
 .desktop-capturer-selection__thumbnail {margin: 5px;width: 150px;}
 .desktop-capturer-selection__name {margin: 10px;overflow-wrap: anywhere;}
 .desktop-capturer-selection__btn:hover, .desktop-capturer-selection__btn:focus {
-  background: #fff;color: #000;;
+    background: %FG%;color: %BG%;
 }`
 const customDisplayMedia = frameWindow => new Promise((resolve, reject) => {
+    let setting = "block"
     let settings = {}
     try {
         settings = JSON.parse(fs.readFileSync(settingsFile).toString())
     } catch (e) {
-        // No webview settings configured, assuming the default "ask"
+        // No webview settings configured, assuming the default "block"
     }
-    if (settings?.permissiondisplaycapture !== "ask") {
+    setting = settings.permissiondisplaycapture || setting
+    let settingRule = ""
+    for (const type of ["ask", "block"]) {
+        for (const r of settings[`permissions${type}ed`]?.split(",")) {
+            if (!r.trim() || settingRule) {
+                continue
+            }
+            const [match, ...names] = r.split("~")
+            if (names.find(p => p.endsWith("displaycapture"))) {
+                if (frameWindow.location.href.match(match)) {
+                    settingRule = type
+                }
+            }
+        }
+    }
+    setting = settingRule || setting
+    // Only blocked requests are notified here right away,
+    // because allowed requests are passed to the regular permission system.
+    if (setting !== "ask") {
+        if (settingRule) {
+            ipcRenderer.sendToHost("notify",
+                `Automatic rule for 'displaycapture' activated at '`
+                + `${frameWindow.location.href}' which was blocked`, "perm")
+        } else {
+            ipcRenderer.sendToHost("notify",
+                `Globally blocked 'displaycapture' at `
+                + `'${frameWindow.location.href}'`, "perm")
+        }
         throw new DOMException("Permission denied", "NotAllowedError")
     }
     return new Promise(async () => {
@@ -216,28 +259,57 @@ const customDisplayMedia = frameWindow => new Promise((resolve, reject) => {
                 {"types": ["screen", "window"]})
             const stylingElem = document.createElement("style")
             stylingElem.textContent = displayCaptureStyling
+                .replace(/%FONTSIZE%/g, settings.fontsize || "14")
+                .replace(/%FG%/g, settings.fg || "#eee")
+                .replace(/%BG%/g, settings.bg || "#333")
+                .replace(/%SHADE%/g, "#7777")
             const selectionElem = document.createElement("div")
             selectionElem.classList = "desktop-capturer-selection"
             selectionElem.innerHTML = `
-                    <div class="desktop-capturer-selection__scroller">
-                      <ul class="desktop-capturer-selection__list">
-                        ${sources.map(details => `
-                          <li class="desktop-capturer-selection__item">
-                            <button class="desktop-capturer-selection__btn"
+            <span class="desktop-capturer-selection__scroller">
+                <span class="desktop-capturer-selection__close">X</span>
+                <span class="desktop-capturer-selection__list">
+                    ${sources.map(details => `
+                        <span class="desktop-capturer-selection__item">
+                            <span class="desktop-capturer-selection__btn"
                                 data-id="${details.id}">
-                              <img class="desktop-capturer-selection__thumbnail"
+                            <img class="desktop-capturer-selection__thumbnail"
                                 src="${details.thumbnail.toDataURL()}" />
-                              <span class="desktop-capturer-selection__name">
+                            <span class="desktop-capturer-selection__name">
                                 ${details.name.trim()}</span>
-                            </button>
-                          </li>
-                        `).join("")}
-                      </ul>
-                    </div>
-                  `
+                            </span>
+                        </span>
+                    `).join("")}
+                </span>
+            </span>`
             document.body.appendChild(selectionElem)
             document.head.appendChild(stylingElem)
-            document.querySelectorAll(
+            const closeButtons = [selectionElem, selectionElem.querySelector(
+                ".desktop-capturer-selection__close")]
+            closeButtons.forEach(button => {
+                button.addEventListener("click", e => {
+                    if (e.path.find(el => el?.matches?.(
+                        ".desktop-capturer-selection__btn"))) {
+                        // Also clicked on a display to share, ignore close
+                        return
+                    }
+                    selectionElem.remove()
+                    stylingElem.remove()
+                    if (settingRule) {
+                        ipcRenderer.sendToHost("notify",
+                            `Ask rule for 'displaycapture' activated at '`
+                            + `${frameWindow.location.href}' which was `
+                            + `blocked by user`, "perm")
+                    } else {
+                        ipcRenderer.sendToHost("notify",
+                            `Manually blocked 'displaycapture' at `
+                            + `'${frameWindow.location.href}'`, "perm")
+                    }
+                    throw new DOMException(
+                        "Permission denied", "NotAllowedError")
+                })
+            })
+            selectionElem.querySelectorAll(
                 ".desktop-capturer-selection__btn").forEach(button => {
                 button.addEventListener("click", async () => {
                     try {